+++ CHANGELOG.md
@@ -10,6 +10,7 @@ I know that many can't update to the latest build. The package will continue wor
 * Added: #59 extended echo and print scope: support.function.construct.output
 * Added: extended boolean scope: constant.language.boolean
 * Added: extended magic constant scope: constant.language.magic
+* Added: extended null constant scope to "constant.language.null"
 * Fixed: #56 indentation for empty brackets
 * Fixed: #58 syntax bracket handling for closures
 * Fixed: #61, #62, and #63 operator scopes not matching correctly